requirement.txt
flask
flask-cors
Werkzeug
mysql-connector-python
python-dotenv
PyPDF2
requests
transformers
torch
pyMuPDF
beautifulsoup4
langchain
langchain-google-genai
google-generativeai
langchain-text-splitters
langchain-community
PyPDF2
docx2txt